fix: get conversation id from conversation object in msteams
When the channel is msteams the CONVERSATION_ID is set to an object instead of a string.
There must have been a bug previously where they were storing this id in the user object.
They fixed their code to store it in the coversation object and this updates our code accordingly.